refactor: 디렉토리 변경 및 코드 정리
