+++ lib/api.js
@@ -27,3 +27,11 @@ export async function getUserProfileById(userId) {
   }
   return data.data[0];
 }
+
+export async function getFolderInfoByFolderId(folderId) {
+  const { data } = await axios.get(`${BASE_URL}/folders/${folderId}`);
+  if (!data.data) {
+    throw new Error("폴더 정보 로딩에 실패했습니다.");
+  }
+  return data.data[0];
+}
