+++ pages/_app.tsx
@@ -1,5 +1,5 @@
-import "@/styles/globals.scss";
 import type { AppProps } from "next/app";
+import "@/styles/globals.scss";
 
 export default function App({ Component, pageProps }: AppProps) {
   return <Component {...pageProps} />;
