+++ pages/folder/modal/modalType/deleteFolder/DeleteFolder.module.scss
