--- pages/folder/modal/modalType/folder/shareFolder.tsx/ShareFolder.module.scss
@@ -1,43 +1,0 @@
-.modal {
-  display: flex;
-  justify-content: center;
-  flex-direction: column;
-  width: 80%;
-  margin: 0 auto;
-  gap: 15px;
-}
-
-.closeButton {
-  position: absolute;
-  width: 24px;
-  height: 24px;
-  padding: 0;
-  top: 15px;
-  right: 15px;
-}
-
-.title {
-  margin: 0px;
-  text-align: center;
-}
-
-.folderName {
-  margin-top: 0;
-  margin-bottom: 20px;
-  text-align: center;
-  color: #9fa6b2;
-  font-size: 16px;
-  font-weight: 400;
-}
-
-.shareButton {
-  display: flex;
-  justify-content: center;
-  gap: 30px;
-
-  button {
-    cursor: pointer;
-    border-style: none;
-    background-color: white;
-  }
-}
